Parsing and internal representation of scrollbar-gutter
The scrollbar-gutter CSS property provides control over the presence
of scrollbar gutters (the space which may be reserved by the UA to
display a scrollbar).

The value of scrollbar-gutter is defined by a combination of one or
more keywords ("auto", "stable", "always", "both" and "force").

This property is internally stored as an unsigned int with binary
flags for each one of those keywords.

For convenience, ScrollbarGutterIs* methods have been added to
ComputedStyle.

Parsing WPT tests are included.

Feature status: test.

Bug: 710214
Change-Id: Ia5928ab4c379097065e28604d8431aab3e8f22c0