Support ServiceWorker interception for requests made from ReadableStream
Before this CL blink::mojom::FetchAPIRequestBody was typemapped to
blink::EncodedFormData. This was not good for a request body made from
ReadableStream because the body is implemented as
blink::ResourceRequestBody::stream_body_ which is out of
EncodedFormData.

Hence this CL changes the type mapping. Now
blink::mojom::FetchAPIRequestBody is typemapped to
blink::ResourceRequestBody.

 - encoded_form_data_mojom_traits.{cc, h} are moved to
   platform/loader/fetch/fetch_api_request_body_mojom_trauts.{cc, h}.
 - encoded_form_data_element_mojom_traits.cc is merged into
   platform/loader/fetch/fetch_api_request_body_mojom_trauts.cc.
   blink.mojom.FetchAPIDataElement is no longer typemapped.
 - Add DataPipeBytesConsumer::CompletionNotifier::SignalSize so that
   it can work with a data pipe coming from ChunkedDataPipeGetter.
 - Support streaming body in FetchRequestData::Create.

Bug: 688906
Change-Id: I8c4ed0f46946aff3f59733dd2138017ead724575